+++ src/cli.ts
@@ -31,6 +31,14 @@ const conf = yargs(process.argv.slice(2))
         type: "string",
         description: "Suffix for generated interface names",
     })
+    .option("caseInsensitiveNames", {
+        type: "boolean",
+        description: "Case-insensitive name while parsing definition names"
+    })
+    .option("maxRecursiveDefinitionName", {
+        type: "number",
+        description: "Maximum count of definition's with same name but increased suffix. Will throw an error if exceed",
+    })
     .option("quiet", {
         type: "boolean",
         description: "Suppress logs",
@@ -57,6 +65,7 @@ if (conf.verbose || process.env.DEBUG) {
 if (conf.quiet) {
     Logger.isDebug = false;
     Logger.isInfo = false;
+    Logger.isWarn = false;
     Logger.isError = false;
 }
 
@@ -75,6 +84,15 @@ if (conf.modelNamePreffix) {
 if (conf.modelNameSuffix) {
     options.modelNameSuffix = conf.modelNameSuffix;
 }
+
+if (conf.maxRecursiveDefinitionName || conf.maxRecursiveDefinitionName == 0) {
+    options.maxRecursiveDefinitionName = conf.maxRecursiveDefinitionName;
+}
+
+if (conf.caseInsensitiveNames) {
+    options.caseInsensitiveNames = conf.caseInsensitiveNames;
+}
+
 Logger.debug("Options");
 Logger.debug(JSON.stringify(options, null, 2));