+++ src/models/parsed-wsdl.ts
@@ -1,4 +1,5 @@
 import sanitizeFilename from "sanitize-filename";
+import { Logger } from "../utils/logger";
 
 export type DefinitionProperty =
     | {
@@ -14,8 +15,8 @@ export type DefinitionProperty =
           sourceName: string;
           description?: string;
           /**
-           * This is very information to know, because
-           * you can avoid circular references with this
+           * This definition only reference another definition instead of primitive type
+           * @description helps to avoid circular referencies
            */
           kind: "REFERENCE";
           isArray?: boolean;
@@ -23,7 +24,9 @@ export type DefinitionProperty =
       };
 
 export interface Definition {
+    /** Will be used as name of generated Definition's interface */
     name: string;
+    /** Original name of Definition in WSDL */
     sourceName: string;
     description?: string;
     docs: string[];
@@ -31,27 +34,47 @@ export interface Definition {
 }
 
 export interface Method {
+    /** Will be used as name for generated Function/Method */
     name: string;
+    /** First param name (InputMessage) */
     paramName: string;
+    /** First param type (InputMessage) */
     paramDefinition: null | Definition;
+    /** Result type (OutputMessage) */
     returnDefinition: null | Definition;
 }
 
 export interface Port {
+    /** Will be used as name of generated Port's interface */
     name: string;
+    /** Original name of Port in WSDL */
     sourceName: string;
     description?: string;
+    /** List of callable methods within this Port */
     methods: Array<Method>;
 }
 
 export interface Service {
+    /** Will be used as name of generated Service's interface */
     name: string;
+    /** Original name of Service in WSDL */
     sourceName: string;
     description?: string;
+    /** List of Service's Ports */
     ports: Array<Port>;
 }
 
-const MAX_STACK = 30;
+export interface Options {
+    caseInsensitiveNames: boolean; 
+    maxStack: number;
+    maxStackWarn: number;
+}
+
+const defaultOptions: Options = {
+    caseInsensitiveNames: false,
+    maxStack: 64,
+    maxStackWarn: 32
+}
 
 export class ParsedWsdl {
     /**
@@ -69,22 +92,43 @@ export class ParsedWsdl {
     ports: Array<Port> = [];
     services: Array<Service> = [];
 
+    private _options: Options;
+    private _warns: string[];
+
+    constructor(options: Partial<Options>) {
+        this._options = {
+            ...defaultOptions,
+            ...options
+        };
+        this._warns = [];
+    }
+
     /** Find definition by it's name */
     findDefinition(definitionName: string): Definition {
         return this.definitions.find((def) => def.name === definitionName);
     }
 
-    /** To make every definition's name unique. If definition with same name exists, suffix it with incremented number */
+    /**
+     * To make every definition's name unique.
+     * If definition with same name exists, suffix it with incremented number
+     */
     findNonCollisionDefinitionName(defName: string): string {
         const definitionName = sanitizeFilename(defName);
-        if (!this.definitions.find((def) => def.name === definitionName)) {
+        const isInSensitive = this._options.caseInsensitiveNames;
+
+        const defNameToCheck = isInSensitive ? definitionName.toLowerCase() : definitionName;
+        if (!this.definitions.find((def) => isInSensitive ? def.name.toLowerCase() : def.name === defNameToCheck)) {
             return definitionName;
         }
-        for (let i = 1; i < MAX_STACK; i++) {
-            if (!this.definitions.find((def) => def.name === `${definitionName}${i}`)) {
+        for (let i = 1; i < this._options.maxStack; i++) {
+            if (!this.definitions.find((def) => isInSensitive ? def.name.toLowerCase() : def.name === `${defNameToCheck}${i}`)) {
                 return `${definitionName}${i}`;
             }
+            if (i == this._options.maxStackWarn && !this._warns.includes(definitionName)) {
+                Logger.warn(`Too many definition with same name "${definitionName}"`);
+                this._warns.push(definitionName);
+            }
         }
-        throw new Error(`Out of stack (${MAX_STACK}) for "${definitionName}", there's probably cyclic definition`);
+        throw new Error(`Out of stack (${this._options.maxStack}) for "${definitionName}", there's probably cyclic definition. You can also try to increase maxStack with --TODO option`);
     }
 }